+++ src/HearThis/UI/BookButton.cs
@@ -1,7 +1,7 @@
 // --------------------------------------------------------------------------------------------
-#region // Copyright (c) 2021, SIL International. All Rights Reserved.
-// <copyright from='2011' to='2021' company='SIL International'>
-//		Copyright (c) 2021, SIL International. All Rights Reserved.
+#region // Copyright (c) 2022, SIL International. All Rights Reserved.
+// <copyright from='2011' to='2022' company='SIL International'>
+//		Copyright (c) 2022, SIL International. All Rights Reserved.
 //
 //		Distributable under the terms of the MIT License (https://sil.mit-license.org/)
 // </copyright>
@@ -26,7 +26,6 @@ namespace HearThis.UI
 		private int _percentageRecorded;
 
 		protected override bool DisplayLabels => DisplayLabelsWhenPaintingButtons;
-		protected override float LabelFontSize => 8;
 
 		protected override ProblemType WorstProblem => _model.GetWorstProblemInBook();