+++ src/HearThis/UI/ScriptControl.cs
@@ -17,6 +17,7 @@ using System.Drawing.Drawing2D;
 using System.Windows.Forms;
 using HearThis.Properties;
 using HearThis.Script;
+using SIL.ObjectModel;
 using SIL.Windows.Forms.Widgets.Flying;
 
 namespace HearThis.UI
@@ -478,14 +479,13 @@ namespace HearThis.UI
 
 		internal bool BrightenContext => _brightenContext;
 
-		public void SetClauseSeparators(string clauseBreakCharacters)
+		public void SetClauseSeparators(IReadOnlySet<char> clauseBreakCharacters)
 		{
 			// Whenever a new project is set or the project's clause-break settings are changed, this should get called to set the
 			// clause break characters stored in the project settings.
-			string clauseSeparatorCharacters = (clauseBreakCharacters ?? Settings.Default.ClauseBreakCharacters).Replace(" ", string.Empty);
-			List<char> clauseSeparators = new List<char>(clauseSeparatorCharacters.ToCharArray());
-			clauseSeparators.Add(ScriptLine.kLineBreak);
-			ClauseSplitter = new SentenceClauseSplitter(clauseSeparators.ToArray());
+			var setWithLineBreakChar = new HashSet<char>(clauseBreakCharacters);
+			setWithLineBreakChar.Add(ScriptLine.kLineBreak);
+			ClauseSplitter = new SentenceClauseSplitter(new ReadOnlySet<char>(setWithLineBreakChar));
 		}
 
 		public enum Direction